redo/requirements.txt
# Core AI and Machine Learning Libraries
openai>=1.0.0
langgraph>=0.2.0
langchain>=0.2.0

# Data Processing and Analysis - PINNED COMPATIBLE VERSIONS
pandas==2.0.3
numpy==1.26.4
openpyxl>=3.1.0
xlrd>=2.0.0

# Image Processing and OCR - PINNED COMPATIBLE VERSIONS  
Pillow>=10.0.0
pytesseract>=0.3.10
opencv-python==4.9.0.80

# Azure Services
azure-core>=1.29.0
azure-ai-documentintelligence>=1.0.0
azure-ai-formrecognizer>=3.3.0

# MSG File Processing
extract-msg>=0.41.0

# Web Scraping and HTML Processing
beautifulsoup4>=4.11.1,<4.13
lxml>=4.9.0

# Configuration and Utilities
toml>=0.10.0
python-dotenv>=1.0.0

# HTTP and API Libraries
requests>=2.31.0
urllib3>=1.25.4,<2.0.0

# Development and Testing (Optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Additional Utilities
python-dateutil>=2.8.0
pathlib2>=2.3.0

# For Windows compatibility
pywin32>=306; sys_platform == "win32" 